Undocumented Lambda request ID magic!
It took me a long time to figure out how the AWS Lambda runtime magically inserts the Lambda request ID into your log output if you use Logger.  I had to figure this out for myself since it stopped working when I overrode the Logger and started to customize it.  I had to use TracePoint to trace what it was doing in Lambda to find the magic, undocumented global variable that the runtime sets.